boards: alif: ensemble: move and fix mram partitions
The mram partiion layout and size are not board dependent. They are
dependent on the SOC variant and the core(HP/HE) the build is
for. So move the partitions into the core specific device tree include
files. Also, fix the partition sizes so that larger mcuboot compatible
images can be built.

Signed-off-by: Silesh C V <silesh@alifsemi.com>