soc: alif: balletto: set default ROM_START_OFFSET
For MCUboot compatible builds, the ROM_START_OFFSET gets
set as 0x200 by default. But as per the alignment requirements
of the vector table, the vector table gets aligned to 0x800 with
this rom offset configuration. With this binary layout, MCUboot
will not be able to locate the vector table of the application
image as the header size (available in the application header)
will be 0x200 while the image starts at offset 0x800. Fix this
by modifying the default ROM_START_OFFSET for MCUboot compatible
builds.

Note that this fix is already available for Ensemble.

Signed-off-by: Silesh C V <silesh@alifsemi.com>